Ajout MangasController et Modif Service

+++ src/main/java/list/perso/PopCulture/service/MangasService.java
@@ -22,7 +22,7 @@ public class MangasService {
 
         if(id == null)
         {
-            throw new RuntimeException("L'ID d'un mangas ne peut pas être null");
+            throw new RuntimeException("L'ID d'un mangas ne peut pas être null"); //Il faut gérer le cas null ici car c'est une simple lecture.
         }
         return mangasRepository.findById(id).orElse(null);
     }
@@ -31,13 +31,29 @@ public class MangasService {
         return mangasRepository.save(mangas);
     }
 
-    public Mangas updateMangas(Mangas mangas) {
-        return mangasRepository.save(mangas);
+    public Mangas updateMangas(Integer id, Mangas mangasBody) {
+
+    Mangas existing = mangasRepository.findById(id).orElse(null);
+
+    if (existing == null) {
+        return null; // On signale au controller que le manga n'existe pas
     }
 
+    existing.setNom(mangasBody.getNom());
+    existing.setAuteur(mangasBody.getAuteur());
+    existing.setTomeLu(mangasBody.getTomeLu());
+    existing.setTomeTotal(mangasBody.getTomeTotal());
 
-    public void deleteMangasById(Integer id) {
-        mangasRepository.deleteById(id);
+    return mangasRepository.save(existing);
+
+    }
+
+    public boolean deleteMangasById(Integer id) {          // Pas besoin de gerer le cas null ici car c'est une "operation critique" donc Spring le gère tout seul.
+        if (mangasRepository.existsById(id)) {
+            mangasRepository.deleteById(id);
+            return true;
+        }
+            return false;
     }
 
     public void deleteMangas(Mangas mangas) {
